+++ models/Dev/data/checkpoint
@@ -1,2 +1,2 @@
-model_checkpoint_path: "C:\\Users\\skirb\\Documents\\Workspace\\SteGANographer\\models\\Dev\\data\\Dev_train-4000"
-all_model_checkpoint_paths: "C:\\Users\\skirb\\Documents\\Workspace\\SteGANographer\\models\\Dev\\data\\Dev_train-4000"
+model_checkpoint_path: "C:\\Users\\skirb\\Documents\\Workspace\\SteGANographer\\models\\Dev\\data\\Dev_train-6000"
+all_model_checkpoint_paths: "C:\\Users\\skirb\\Documents\\Workspace\\SteGANographer\\models\\Dev\\data\\Dev_train-6000"
